+++ infra/terraform/postgres/outputs.tf
@@ -1,1 +1,0 @@
-output "ip_address" { value = "${ google_sql_database_instance.postgres.ip_address.0.ip_address}" }